+++ RELEASE-NOTES.md
@@ -1,3 +1,7 @@
+## Release 3.8
+
+* #468 ClassCastException in combination Agrest 3.4 and Cayenne 4.1
+
 ## Release 3.7
 
 * #449 Excludes in overlays: support removing attributes and relationships via overlays